smolbench: Report upsert server latencies
