+++ smolbench/src/apis.rs
@@ -1,0 +1,79 @@
+use crate::error::SmolBenchError;
+use crate::types::{ApiResponse, ApiSuccessResponse, Point};
+use http::Uri;
+use serde_json::{json, Value};
+
+pub async fn create_collection(
+    url: &Uri,
+    collection_name: &str,
+) -> Result<ApiSuccessResponse<Value>, SmolBenchError> {
+    let client = reqwest::Client::new();
+
+    let res = client
+        .put(format!("{url}/collections/{collection_name}"))
+        .json(&serde_json::json!({
+            "params": "..."
+        }))
+        .send()
+        .await?;
+
+    let body: ApiResponse<Value> = res.json().await?;
+
+    match body {
+        ApiResponse::Success(body) => {
+            dbg!(&body.result);
+            Ok(body)
+        }
+        ApiResponse::Error(res) => Err(SmolBenchError::CreateCollectionError(res.error)),
+    }
+}
+
+pub async fn upsert_points(
+    url: &Uri,
+    collection_name: &str,
+    num_points: usize,
+    batch_size: usize,
+) -> Result<Vec<ApiSuccessResponse<Value>>, SmolBenchError> {
+    let client = reqwest::Client::new();
+    let num_batches = num_points.div_ceil(batch_size);
+
+    let mut results = Vec::with_capacity(num_batches);
+
+    for batch in 0..num_batches {
+        let start = batch * batch_size;
+        let end = std::cmp::min(start + batch_size, num_points);
+
+        let batch_ts = chrono::Utc::now();
+
+        let points: Vec<Point> = (start..end)
+            .map(|i| Point {
+                id: i,
+                payload: json!({
+                    "text": format!("Point {}", i),
+                    "timestamp": batch_ts.to_rfc3339(),
+                }),
+            })
+            .collect();
+
+        let res = client
+            .put(format!("{url}/collections/{collection_name}/points"))
+            .json(&json!({
+                "points": points,
+            }))
+            .send()
+            .await?;
+
+        let text = res.text().await?;
+
+        let body: ApiResponse<Value> = serde_json::from_str(&text)?;
+
+        match body {
+            ApiResponse::Success(body) => results.push(body),
+            ApiResponse::Error(res) => {
+                return Err(SmolBenchError::CreateCollectionError(res.error));
+            }
+        }
+    }
+
+    Ok(results)
+}
