+++ smolbench/src/main.rs
@@ -1,87 +1,11 @@
+pub mod apis;
 pub mod args;
 pub mod error;
 pub mod types;
 
-use crate::types::{ApiResponse, ApiSuccessResponse, Point};
+use crate::apis::{create_collection, upsert_points};
 use args::parse_args;
 use error::SmolBenchError;
-use http::Uri;
-use serde_json::{json, Value};
-
-async fn create_collection(
-    url: &Uri,
-    collection_name: &str,
-) -> Result<ApiSuccessResponse<Value>, SmolBenchError> {
-    let client = reqwest::Client::new();
-
-    let res = client
-        .put(format!("{url}/collections/{collection_name}"))
-        .json(&serde_json::json!({
-            "params": "..."
-        }))
-        .send()
-        .await?;
-
-    let body: ApiResponse<Value> = res.json().await?;
-
-    match body {
-        ApiResponse::Success(body) => {
-            dbg!(&body.result);
-            Ok(body)
-        }
-        ApiResponse::Error(res) => Err(SmolBenchError::CreateCollectionError(res.error)),
-    }
-}
-
-async fn upsert_points(
-    url: &Uri,
-    collection_name: &str,
-    num_points: usize,
-    batch_size: usize,
-) -> Result<Vec<ApiSuccessResponse<Value>>, SmolBenchError> {
-    let client = reqwest::Client::new();
-    let num_batches = num_points.div_ceil(batch_size);
-
-    let mut results = Vec::with_capacity(num_batches);
-
-    for batch in 0..num_batches {
-        let start = batch * batch_size;
-        let end = std::cmp::min(start + batch_size, num_points);
-
-        let batch_ts = chrono::Utc::now();
-
-        let points: Vec<Point> = (start..end)
-            .map(|i| Point {
-                id: i,
-                payload: json!({
-                    "text": format!("Point {}", i),
-                    "timestamp": batch_ts.to_rfc3339(),
-                }),
-            })
-            .collect();
-
-        let res = client
-            .put(format!("{url}/collections/{collection_name}/points"))
-            .json(&json!({
-                "points": points,
-            }))
-            .send()
-            .await?;
-
-        let text = res.text().await?;
-
-        let body: ApiResponse<Value> = serde_json::from_str(&text)?;
-
-        match body {
-            ApiResponse::Success(body) => results.push(body),
-            ApiResponse::Error(res) => {
-                return Err(SmolBenchError::CreateCollectionError(res.error));
-            }
-        }
-    }
-
-    Ok(results)
-}
 
 #[tokio::main]
 async fn main() -> Result<(), SmolBenchError> {
@@ -112,9 +36,9 @@ async fn main() -> Result<(), SmolBenchError> {
     let min_latency = latencies.iter().cloned().fold(f64::INFINITY, f64::min);
     let max_latency = latencies.iter().cloned().fold(f64::NEG_INFINITY, f64::max);
 
-    println!("Average batch latency: {:.2} ms", avg_latency);
-    println!("Minimum batch latency: {:.2} ms", min_latency);
-    println!("Maximum batch latency: {:.2} ms", max_latency);
+    println!("Avg batch server latency: {:.2} ms", avg_latency);
+    println!("Min batch server latency: {:.2} ms", min_latency);
+    println!("Max batch server latency: {:.2} ms", max_latency);
 
     println!(
         "Upserted {} points in batches of {} into collection '{}'",
